+++ Containerfile.supervisor
@@ -49,6 +49,7 @@ RUN useradd -m -G wheel beeai
 # Copy required directories
 # Individual directories are mounted as volumes in development
 COPY agents/ /home/beeai/agents/
+COPY ymir_tools/ /home/beeai/ymir_tools/
 COPY common/ /home/beeai/common/
 COPY supervisor/ /home/beeai/supervisor/
 RUN chgrp -R root /home/beeai && chmod -R g+rX /home/beeai
