+++ Makefile.tests
@@ -20,7 +20,7 @@ check-agents:
 	cd ./agents && \
 	PYTHONPATH=$(CURDIR) PYTHONDONTWRITEBYTECODE=1 python3 -m pytest --verbose --showlocals $(TEST_TARGET)
 check-mcp-server:
-	cd ./mcp_server && \
+	cd ./ymir_tools/privileged && \
 	PYTHONPATH=$(CURDIR) PYTHONDONTWRITEBYTECODE=1 python3 -m pytest --verbose --showlocals $(TEST_TARGET)
 check-jira-issue-fetcher:
 	cd ./jira_issue_fetcher && \
