+++ agents/merge_request_agent.py
@@ -36,9 +36,9 @@ from common.models import (
 from common.utils import is_cs_branch
 from constants import I_AM_YMIR
 from observability import setup_observability
-from tools.commands import RunShellCommandTool
-from tools.filesystem import GetCWDTool, RemoveTool
-from tools.text import (
+from ymir_tools.unprivileged.commands import RunShellCommandTool
+from ymir_tools.unprivileged.filesystem import GetCWDTool, RemoveTool
+from ymir_tools.unprivileged.text import (
     CreateTool,
     InsertAfterSubstringTool,
     InsertTool,
