+++ agents/tests/unit/test_zstream_search.py
@@ -5,7 +5,7 @@ import pytest
 
 from beeai_framework.middleware.trajectory import GlobalTrajectoryMiddleware
 
-from tools.zstream_search import (
+from ymir_tools.unprivileged.zstream_search import (
     ZStreamSearchTool,
     ZStreamSearchToolInput,
     ZStreamSearchResult,
@@ -138,7 +138,7 @@ async def test_not_applicable_y_stream():
 async def test_not_applicable_current_zstream():
     """Current z-stream fixVersion should return NOT_APPLICABLE."""
     tool = ZStreamSearchTool()
-    with patch("tools.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=False):
+    with patch("ymir_tools.unprivileged.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=False):
         output = await tool.run(
             input=ZStreamSearchToolInput(
                 component="fence-agents",
@@ -189,9 +189,9 @@ async def test_found_in_closest_stream():
 
     tool = ZStreamSearchTool()
     with (
-        patch("tools.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=True),
-        patch("tools.zstream_search.mcp_tools", mock_mcp_tools),
-        patch("tools.zstream_search.run_tool", mock_run_tool),
+        patch("ymir_tools.unprivileged.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=True),
+        patch("ymir_tools.unprivileged.zstream_search.mcp_tools", mock_mcp_tools),
+        patch("ymir_tools.unprivileged.zstream_search.run_tool", mock_run_tool),
         patch.dict("os.environ", {"MCP_GATEWAY_URL": "http://mcp-gateway:8000"}),
     ):
         output = await tool.run(
@@ -241,9 +241,9 @@ async def test_cascade_to_further_version():
 
     tool = ZStreamSearchTool()
     with (
-        patch("tools.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=True),
-        patch("tools.zstream_search.mcp_tools", mock_mcp_tools),
-        patch("tools.zstream_search.run_tool", mock_run_tool),
+        patch("ymir_tools.unprivileged.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=True),
+        patch("ymir_tools.unprivileged.zstream_search.mcp_tools", mock_mcp_tools),
+        patch("ymir_tools.unprivileged.zstream_search.run_tool", mock_run_tool),
         patch.dict("os.environ", {"MCP_GATEWAY_URL": "http://mcp-gateway:8000"}),
     ):
         output = await tool.run(
@@ -277,9 +277,9 @@ async def test_not_found_anywhere():
 
     tool = ZStreamSearchTool()
     with (
-        patch("tools.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=True),
-        patch("tools.zstream_search.mcp_tools", mock_mcp_tools),
-        patch("tools.zstream_search.run_tool", mock_run_tool),
+        patch("ymir_tools.unprivileged.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=True),
+        patch("ymir_tools.unprivileged.zstream_search.mcp_tools", mock_mcp_tools),
+        patch("ymir_tools.unprivileged.zstream_search.run_tool", mock_run_tool),
         patch.dict("os.environ", {"MCP_GATEWAY_URL": "http://mcp-gateway:8000"}),
     ):
         output = await tool.run(
@@ -302,9 +302,9 @@ async def test_no_related_issues_found():
 
     tool = ZStreamSearchTool()
     with (
-        patch("tools.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=True),
-        patch("tools.zstream_search.mcp_tools", mock_mcp_tools),
-        patch("tools.zstream_search.run_tool", mock_run_tool),
+        patch("ymir_tools.unprivileged.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=True),
+        patch("ymir_tools.unprivileged.zstream_search.mcp_tools", mock_mcp_tools),
+        patch("ymir_tools.unprivileged.zstream_search.run_tool", mock_run_tool),
         patch.dict("os.environ", {"MCP_GATEWAY_URL": "http://mcp-gateway:8000"}),
     ):
         output = await tool.run(
@@ -350,9 +350,9 @@ async def test_version_proximity_sorting():
 
     tool = ZStreamSearchTool()
     with (
-        patch("tools.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=True),
-        patch("tools.zstream_search.mcp_tools", mock_mcp_tools),
-        patch("tools.zstream_search.run_tool", mock_run_tool),
+        patch("ymir_tools.unprivileged.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=True),
+        patch("ymir_tools.unprivileged.zstream_search.mcp_tools", mock_mcp_tools),
+        patch("ymir_tools.unprivileged.zstream_search.run_tool", mock_run_tool),
         patch.dict("os.environ", {"MCP_GATEWAY_URL": "http://mcp-gateway:8000"}),
     ):
         output = await tool.run(
@@ -393,9 +393,9 @@ async def test_older_issues_excluded():
 
     tool = ZStreamSearchTool()
     with (
-        patch("tools.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=True),
-        patch("tools.zstream_search.mcp_tools", mock_mcp_tools),
-        patch("tools.zstream_search.run_tool", mock_run_tool),
+        patch("ymir_tools.unprivileged.zstream_search.is_older_zstream", new_callable=AsyncMock, return_value=True),
+        patch("ymir_tools.unprivileged.zstream_search.mcp_tools", mock_mcp_tools),
+        patch("ymir_tools.unprivileged.zstream_search.run_tool", mock_run_tool),
         patch.dict("os.environ", {"MCP_GATEWAY_URL": "http://mcp-gateway:8000"}),
     ):
         output = await tool.run(
